Creating a separate tdi_json parsing library (#4)
* Updating target-utils and syslibs refpoints for standalone
* Adding cmake option for standlone build
