+++ .github/workflows/cmake.yml
@@ -29,7 +29,7 @@ jobs:
     - name: Configure CMake
       # Configure CMake in a 'build' subdirectory. `CMAKE_BUILD_TYPE` is only required if you are using a single-configuration generator such as make.
       # See https://cmake.org/cmake/help/latest/variable/CMAKE_BUILD_TYPE.html?highlight=cmake_build_type
-      run: mkdir -p build && cd build && cmake .. -DCMAKE_INSTALL_PREFIX=${{github.workspace}}/install
+      run: mkdir -p build && cd build && cmake .. -DSTANDALONE=on -DCMAKE_INSTALL_PREFIX=${{github.workspace}}/install
 
     - name: Build
       # Build your program with the given configuration
