+++ include/tdi/common/tdi_table.hpp
@@ -29,7 +29,7 @@
 #include <unordered_map>
 
 #include <tdi/common/tdi_defs.h>
-#include <tdi/common/tdi_table_info.hpp>
+//#include <tdi/common/tdi_table_info.hpp>
 #include <tdi/common/tdi_target.hpp>
 #include <tdi/common/tdi_session.hpp>
 #include <tdi/common/tdi_table_data.hpp>
